DataBase

+++ app/handlers/commands.py
@@ -1,25 +1,31 @@
 from aiogram import types
 
 from app.handlers.quiz import users
-from main import dp, bot
+from app.utility.text_end import count_text
+from main import dp, bot, users_db
 
 
 # обработчик команды start
-@dp.message_handler(commands='start')
+@dp.message_handler(commands=['start'])
 async def cmd_start(message: types.Message):
     """В ответ на команду start выводит сообщение о начале игры и соответствующую кнопку и обнуляет счетчики вопросов и
      выйгрыша"""
-    await bot.delete_message(chat_id=message.from_user.id, message_id=message.message_id - 1)
-    keyboard = types.InlineKeyboardMarkup()
-    keyboard.add(types.InlineKeyboardButton(text='Начать игру', callback_data='start_quiz'))
-    await bot.send_message(chat_id=message.from_user.id,
-                           text='Добро пожаловать на игру: "Кто хочет стать милионнером?"\n\n'
-                                'Вам необходимо ответить на 15 вопросов, которые разделенны на 4 категории сложности. '
-                                'Ответив на все 15 вопросов Вы получите 1 000 000 рублей. Ответив неверно, Вы '
-                                'проиграете и останетесь с несгораемой суммой до которой успели дойти.\n',
-                           reply_markup=keyboard)
+    user_ref = 0 if len(message.text.split()) < 2 else message.text.split()[1]
+    if users_db.add_new_user(user_id=message.from_user.id, user_name=message.from_user.first_name, user_ref=user_ref):
+        await bot.delete_message(chat_id=message.from_user.id, message_id=message.message_id - 1)
+        keyboard = types.InlineKeyboardMarkup()
+        keyboard.add(types.InlineKeyboardButton(text='Начать игру', callback_data='start_quiz'))
+        await bot.send_message(chat_id=message.from_user.id,
+                               text='Добро пожаловать на игру: "Кто хочет стать милионнером?"\n\n'
+                                    'Вам необходимо ответить на 15 вопросов, которые разделенны на 4 категории '
+                                    'сложности. '
+                                    'Ответив на все 15 вопросов Вы получите 1 000 000 рублей. Ответив неверно, Вы '
+                                    'проиграете и останетесь с несгораемой суммой до которой успели дойти.\n',
+                               reply_markup=keyboard)
 
-    users.clear()
+        users.clear()
+        return
+    return await message.answer("Вы уже были")
 
 
 # обработчик команды help
@@ -43,3 +49,12 @@ async def cmd_exit(message: types.message):
     """В ответ на команду exit обнуляет счетчики вопросов и выйгрыша и завершает игру"""
     await bot.delete_message(chat_id=message.from_user.id, message_id=message.message_id - 1)
     users.clear()
+
+
+# реферальная ссылка
+@dp.message_handler(commands=['ref'])
+async def ref_funk(message: types.Message):
+    count_users = users_db.count_ref(message.from_user.id)
+    refs = f'Вы пригласили {count_users} пользовател' + count_text(count_users, ['я', 'eй', 'ей'])
+    link = 'https://t.me/imsr_su_bot?start=' + str(message.from_user.id)
+    return await message.answer(refs + '\nВаша реф ссылка ' + link)
